+++ app/src/main/java/fr/delcey/pokedexino/data/pokemons/pokeapi/pokemon/GameIndicesItem.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class GameIndicesItem(
 
     @field:SerializedName("game_index")
-    val gameIndex: Int? = null,
+    val gameIndex: Long? = null,
 
     @field:SerializedName("version")
     val version: Version? = null
