+++ app/src/main/java/fr/delcey/pokedexino/data/pokemons/pokeapi/pokemon/TypesItem.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class TypesItem(
 
     @field:SerializedName("slot")
-    val slot: Int? = null,
+    val slot: Long? = null,
 
     @field:SerializedName("type")
     val type: Type? = null
